AD/private/Moda_sploit/requirements.txt
asn1tools==0.155.3
bitstruct==8.11.1
cffi==1.14.4
cryptography==3.3.1
diskcache==5.2.1
gmpy2==2.1.2
prompt-toolkit==3.0.14
pycparser==2.20
pycrypto==2.6.1
pyparsing==2.4.7
six==1.15.0
wcwidth==0.2.5
requests